requirements.txt
pytesseract==0.3.10
opencv-python==4.8.0.76
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
googletrans==3.1.0a0
Pillow==10.0.0
tensorflow==2.13.0
keras==2.13.1
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
google-generativeai>=0.3.0
gTTS>=2.5.0
playsound>=1.3.0 